Refactoring + test
